+++ src/main/java/com/JsonAjax/justcompiler/Syntax/ExpressionSyntax.java
@@ -1,0 +1,14 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package com.JsonAjax.justcompiler.Syntax;
+
+/**
+ *
+ * @author ajax
+ */
+public abstract class ExpressionSyntax extends SyntaxNode {
+    
+}
