+++ src/main/java/com/JsonAjax/justcompiler/Syntax/Lexer.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.JsonAjax.justcompiler;
+package com.JsonAjax.justcompiler.Syntax;
 
 import java.util.ArrayList;
