+++ src/main/java/com/JsonAjax/justcompiler/Syntax/LiteralExpressionSyntax.java
@@ -3,10 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.JsonAjax.justcompiler.Parcing;
-
-import com.JsonAjax.justcompiler.SyntaxKind;
-import com.JsonAjax.justcompiler.SyntaxToken;
+package com.JsonAjax.justcompiler.Syntax;
 
 /**
  *
